refactor blobfuse2
