+++ pkg/apis/utils/labels/labels_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/equinor/radix-operator/pkg/apis/kube"
-	"github.com/equinor/radix-operator/pkg/apis/radix/v1"
+	v1 "github.com/equinor/radix-operator/pkg/apis/radix/v1"
 	"github.com/stretchr/testify/assert"
 	kubelabels "k8s.io/apimachinery/pkg/labels"
 )
@@ -178,6 +178,17 @@ func Test_ForExternalDNSTLSSecret(t *testing.T) {
 func Test_ForExternalDNSCertificate(t *testing.T) {
 	actual := ForExternalDNSCertificate("any-app", v1.RadixDeployExternalDNS{FQDN: "test.com"})
 	expected := kubelabels.Set{kube.RadixAppLabel: "any-app", kube.RadixExternalAliasFQDNLabel: "test.com"}
+	assert.Equal(t, expected, actual)
+}
+
+func Test_ForBlobCSIAzurePersistentVolume(t *testing.T) {
+	actual := ForBlobCSIAzurePersistentVolume("any-app", "any-ns", "any-comp", v1.RadixVolumeMount{Name: "any-vol"})
+	expected := kubelabels.Set{
+		kube.RadixAppLabel:             "any-app",
+		kube.RadixNamespace:            "any-ns",
+		kube.RadixComponentLabel:       "any-comp",
+		kube.RadixVolumeMountNameLabel: "any-vol",
+	}
 	assert.Equal(t, expected, actual)
 }
