+++ pkg/apis/volumemount/volumemount.go
@@ -18,7 +18,6 @@ import (
 	"github.com/rs/zerolog/log"
 	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 )
@@ -66,6 +65,19 @@ func GetVolumes(ctx context.Context, kubeUtil *kube.Kube, namespace string, depl
 	volumes = append(volumes, storageRefsVolumes...)
 
 	return volumes, nil
+}
+
+// GetExistingJobAuxComponentVolumes Get existing job aux component volumes
+func GetExistingJobAuxComponentVolumes(ctx context.Context, kubeUtil *kube.Kube, namespace, jobComponentName string) ([]corev1.Volume, error) {
+	jobAuxKubeDeploymentName := defaults.GetJobAuxKubeDeployName(jobComponentName)
+	jobAuxKubeDeployment, err := kubeUtil.KubeClient().AppsV1().Deployments(namespace).Get(ctx, jobAuxKubeDeploymentName, metav1.GetOptions{})
+	if err != nil {
+		if k8serrors.IsNotFound(err) {
+			return nil, nil
+		}
+		return nil, err
+	}
+	return jobAuxKubeDeployment.Spec.Template.Spec.Volumes, nil
 }
 
 // GarbageCollectVolumeMountsSecretsNoLongerInSpecForComponent Garbage collect volume-mount related secrets that are no longer in the spec
@@ -123,38 +135,6 @@ func CreateOrUpdateVolumeMountSecrets(ctx context.Context, kubeUtil *kube.Kube, 
 	return volumeMountSecretsToManage, nil
 }
 
-// GetCsiAzureVolumeMountType Gets the CSI Azure volume mount type
-func GetCsiAzureVolumeMountType(radixVolumeMount *radixv1.RadixVolumeMount) radixv1.MountType {
-	if radixVolumeMount.BlobFuse2 == nil {
-		//nolint:staticcheck
-		return radixVolumeMount.Type
-	}
-	switch radixVolumeMount.BlobFuse2.Protocol {
-	case radixv1.BlobFuse2ProtocolFuse2, "": // default protocol if not set
-		return radixv1.MountTypeBlobFuse2Fuse2CsiAzure
-	default:
-		return "unsupported"
-	}
-}
-
-// CreateOrUpdateCsiAzureVolumeResourcesForDeployComponent Create or update CSI Azure volume resources for a DeployComponent - PersistentVolumes, PersistentVolumeClaims, PersistentVolume
-// Returns actual volumes, with existing relevant PersistentVolumeClaimName and PersistentVolumeName
-func CreateOrUpdateCsiAzureVolumeResourcesForDeployComponent(ctx context.Context, kubeClient kubernetes.Interface, radixDeployment *radixv1.RadixDeployment, namespace string, deployComponent radixv1.RadixCommonDeployComponent, desiredVolumes []corev1.Volume) ([]corev1.Volume, error) {
-	componentName := deployComponent.GetName()
-	actualVolumes, err := createOrUpdateCsiAzureVolumeResourcesForVolumes(ctx, kubeClient, radixDeployment, namespace, componentName, deployComponent.GetIdentity(), desiredVolumes)
-	if err != nil {
-		return nil, err
-	}
-	return actualVolumes, nil
-}
-
-type blobfuse2VolumeResourceHandler struct {
-	kubeClient      kubernetes.Interface
-	radixDeployment *radixv1.RadixDeployment
-	deployComponent radixv1.RadixCommonDeployComponent
-	desiredVolumes  []corev1.Volume
-}
-
 func getCsiAzurePvsForNamespace(ctx context.Context, kubeClient kubernetes.Interface, namespace string, onlyFunctional bool) ([]corev1.PersistentVolume, error) {
 	pvList, err := kubeClient.CoreV1().PersistentVolumes().List(ctx, metav1.ListOptions{})
 	if err != nil {
@@ -163,14 +143,6 @@ func getCsiAzurePvsForNamespace(ctx context.Context, kubeClient kubernetes.Inter
 	return slice.FindAll(pvList.Items, func(pv corev1.PersistentVolume) bool {
 		return pvIsForCsiDriver(pv) && pvIsForNamespace(pv, namespace) && (!onlyFunctional || pvIsFunctional(pv))
 	}), nil
-}
-
-func isKnownCsiAzureVolumeMount(volumeMount string) bool {
-	switch volumeMount {
-	case string(radixv1.MountTypeBlobFuse2FuseCsiAzure), string(radixv1.MountTypeBlobFuse2Fuse2CsiAzure):
-		return true
-	}
-	return false
 }
 
 func getRadixComponentVolumeMounts(deployComponent radixv1.RadixCommonDeployComponent) ([]corev1.VolumeMount, error) {
@@ -228,13 +200,11 @@ func getCsiAzureVolumeMountName(componentName string, volumeMount *radixv1.Radix
 	if len(volumeMount.Name) == 0 {
 		return "", fmt.Errorf("name is empty for volume mount in the component %s", componentName)
 	}
-	csiAzureVolumeStorageName := getRadixVolumeMountStorage(volumeMount)
+	csiAzureVolumeStorageName := volumeMount.GetStorageContainerName()
 	if len(csiAzureVolumeStorageName) == 0 {
 		return "", fmt.Errorf("storage is empty for volume mount %s in the component %s", volumeMount.Name, componentName)
 	}
-	if len(volumeMount.Path) == 0 {
-		return "", fmt.Errorf("path is empty for volume mount %s in the component %s", volumeMount.Name, componentName)
-	}
+
 	return trimVolumeNameToValidLength(fmt.Sprintf(csiVolumeNameTemplate, csiVolumeType, componentName, volumeMount.Name, csiAzureVolumeStorageName)), nil
 }
 
@@ -437,7 +407,7 @@ func createOrUpdateCsiAzureVolumeMountsSecrets(ctx context.Context, kubeUtil *ku
 			Labels: map[string]string{
 				kube.RadixAppLabel:             appName,
 				kube.RadixComponentLabel:       componentName,
-				kube.RadixMountTypeLabel:       string(GetCsiAzureVolumeMountType(radixVolumeMount)),
+				kube.RadixMountTypeLabel:       string(radixVolumeMount.GetVolumeMountType()),
 				kube.RadixVolumeMountNameLabel: radixVolumeMount.Name,
 			},
 		},
@@ -477,213 +447,10 @@ func getCsiAzurePvcs(ctx context.Context, kubeClient kubernetes.Interface, names
 	return kubeClient.CoreV1().PersistentVolumeClaims(namespace).List(ctx, metav1.ListOptions{})
 }
 
-func buildPvc(appName, namespace, componentName, pvName, pvcName string, radixVolumeMount *radixv1.RadixVolumeMount) *corev1.PersistentVolumeClaim {
-	return &corev1.PersistentVolumeClaim{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pvcName,
-			Namespace: namespace,
-			Labels: map[string]string{
-				kube.RadixAppLabel:             appName,
-				kube.RadixComponentLabel:       componentName,
-				kube.RadixMountTypeLabel:       string(GetCsiAzureVolumeMountType(radixVolumeMount)),
-				kube.RadixVolumeMountNameLabel: radixVolumeMount.Name,
-			},
-		},
-		Spec: corev1.PersistentVolumeClaimSpec{
-			AccessModes: []corev1.PersistentVolumeAccessMode{getVolumeMountAccessMode(radixVolumeMount)},
-			Resources: corev1.VolumeResourceRequirements{
-				Requests: corev1.ResourceList{corev1.ResourceStorage: getVolumeCapacity(radixVolumeMount)},
-			},
-			VolumeName:       pvName,
-			StorageClassName: pointers.Ptr(""), // use "" to avoid to use the "default" storage class
-			VolumeMode:       pointers.Ptr(corev1.PersistentVolumeFilesystem),
-		},
-	}
-}
-
-func getVolumeCapacity(radixVolumeMount *radixv1.RadixVolumeMount) resource.Quantity {
-	requestsVolumeMountSize, err := resource.ParseQuantity(getRadixBlobFuse2VolumeMountRequestsStorage(radixVolumeMount))
-	if err != nil {
-		return resource.MustParse("1Mi")
-	}
-	return requestsVolumeMountSize
-}
-
-func buildCsiAzurePv(appName, namespace, componentName, pvName, pvcName string, radixVolumeMount *radixv1.RadixVolumeMount, identity *radixv1.Identity) *corev1.PersistentVolume {
-	identityClientId := identity.GetAzure().GetClientId()
-	pv := corev1.PersistentVolume{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:   pvName,
-			Labels: getCsiAzurePvLabels(appName, namespace, componentName, radixVolumeMount),
-		},
-		Spec: corev1.PersistentVolumeSpec{
-			StorageClassName: "",
-			MountOptions:     getCsiAzurePvMountOptionsForAzureBlob(radixVolumeMount),
-			Capacity:         corev1.ResourceList{corev1.ResourceStorage: getVolumeCapacity(radixVolumeMount)},
-			AccessModes:      []corev1.PersistentVolumeAccessMode{getVolumeMountAccessMode(radixVolumeMount)},
-			ClaimRef: &corev1.ObjectReference{
-				APIVersion: "v1",
-				Kind:       k8s.KindPersistentVolumeClaim,
-				Namespace:  namespace,
-				Name:       pvcName,
-			},
-			PersistentVolumeSource: corev1.PersistentVolumeSource{
-				CSI: &corev1.CSIPersistentVolumeSource{
-					Driver:           provisionerBlobCsiAzure,
-					VolumeHandle:     getVolumeHandle(namespace, componentName, pvName, getRadixVolumeMountStorage(radixVolumeMount)),
-					VolumeAttributes: getCsiAzurePvAttributes(namespace, radixVolumeMount, identityClientId),
-				},
-			},
-		},
-	}
-	if !radixVolumeMount.UseAzureIdentity() {
-		csiVolumeCredSecretName := defaults.GetCsiAzureVolumeMountCredsSecretName(componentName, radixVolumeMount.Name)
-		pv.Spec.CSI.NodeStageSecretRef = &corev1.SecretReference{Name: csiVolumeCredSecretName, Namespace: namespace}
-	}
-	pv.Spec.PersistentVolumeReclaimPolicy = corev1.PersistentVolumeReclaimRetain // Using only PersistentVolumeReclaimRetain. PersistentVolumeReclaimPolicy deletes volume on unmount.
-	return &pv
-}
-
 func getVolumeHandle(namespace, componentName, pvName, storageName string) string {
 	// Specify a value the driver can use to uniquely identify the share in the cluster.
 	// https://github.com/kubernetes-csi/csi-driver-smb/blob/master/docs/driver-parameters.md#pvpvc-usage
 	return fmt.Sprintf("%s#%s#%s#%s", namespace, componentName, pvName, storageName)
-}
-
-func getCsiAzurePvLabels(appName, namespace, componentName string, radixVolumeMount *radixv1.RadixVolumeMount) map[string]string {
-	return map[string]string{
-		kube.RadixAppLabel:             appName,
-		kube.RadixNamespace:            namespace,
-		kube.RadixComponentLabel:       componentName,
-		kube.RadixVolumeMountNameLabel: radixVolumeMount.Name,
-	}
-}
-
-func getCsiAzurePvAttributes(namespace string, radixVolumeMount *radixv1.RadixVolumeMount, clientId string) map[string]string {
-	attributes := make(map[string]string)
-	switch GetCsiAzureVolumeMountType(radixVolumeMount) {
-	case radixv1.MountTypeBlobFuse2FuseCsiAzure:
-		attributes[csiVolumeMountAttributeContainerName] = getRadixBlobFuse2VolumeMountContainerName(radixVolumeMount)
-		attributes[csiVolumeMountAttributeProtocol] = csiVolumeAttributeProtocolParameterFuse
-		attributes[csiVolumeMountAttributeSecretNamespace] = namespace
-	case radixv1.MountTypeBlobFuse2Fuse2CsiAzure:
-		attributes[csiVolumeMountAttributeContainerName] = getRadixBlobFuse2VolumeMountContainerName(radixVolumeMount)
-		attributes[csiVolumeMountAttributeProtocol] = csiVolumeAttributeProtocolParameterFuse2
-		if radixVolumeMount.BlobFuse2 != nil {
-			if len(radixVolumeMount.BlobFuse2.StorageAccount) > 0 {
-				attributes[csiVolumeAttributeStorageAccount] = radixVolumeMount.BlobFuse2.StorageAccount
-			}
-			if radixVolumeMount.UseAzureIdentity() {
-				attributes[csiVolumeAttributeClientID] = clientId
-				attributes[csiVolumeAttributeResourceGroup] = radixVolumeMount.BlobFuse2.ResourceGroup
-				if len(radixVolumeMount.BlobFuse2.SubscriptionId) > 0 {
-					attributes[csiVolumeAttributeSubscriptionId] = radixVolumeMount.BlobFuse2.SubscriptionId
-				}
-				if len(radixVolumeMount.BlobFuse2.TenantId) > 0 {
-					attributes[csiVolumeAttributeTenantId] = radixVolumeMount.BlobFuse2.TenantId
-				}
-			} else {
-				attributes[csiVolumeMountAttributeSecretNamespace] = namespace
-			}
-		}
-	}
-	// Do not specify the key storage.kubernetes.io/csiProvisionerIdentity in csi.volumeAttributes in PV specification. This key indicates dynamically provisioned PVs
-	// https://github.com/kubernetes-csi/external-provisioner/blob/master/pkg/controller/controller.go#L289C5-L289C21
-	// It looks like this: storage.kubernetes.io/csiProvisionerIdentity: 1731647415428-2825-blob.csi.azure.com
-	return attributes
-}
-
-func getCsiAzurePvMountOptionsForAzureBlob(radixVolumeMount *radixv1.RadixVolumeMount) []string {
-	mountOptions := []string{
-		"--file-cache-timeout-in-seconds=120",
-		"--use-attr-cache=true",
-		"--cancel-list-on-mount-seconds=0",
-		"-o allow_other",
-		"-o attr_timeout=120",
-		"-o entry_timeout=120",
-		"-o negative_timeout=120",
-	}
-	gid := getRadixBlobFuse2VolumeMountGid(radixVolumeMount)
-	if len(gid) > 0 {
-		mountOptions = append(mountOptions, fmt.Sprintf("-o %s=%s", csiMountOptionGid, gid))
-	} else {
-		uid := getRadixBlobFuse2VolumeMountUid(radixVolumeMount)
-		if len(uid) > 0 {
-			mountOptions = append(mountOptions, fmt.Sprintf("-o %s=%s", csiMountOptionUid, uid))
-		}
-	}
-	if getVolumeMountAccessMode(radixVolumeMount) == corev1.ReadOnlyMany {
-		mountOptions = append(mountOptions, "--read-only=true")
-	}
-	if radixVolumeMount.BlobFuse2 != nil {
-		mountOptions = append(mountOptions, getStreamingMountOptions(radixVolumeMount.BlobFuse2.StreamingOptions)...)
-		mountOptions = append(mountOptions, fmt.Sprintf("--%s=%v", csiMountOptionUseAdls, radixVolumeMount.BlobFuse2.UseAdls != nil && *radixVolumeMount.BlobFuse2.UseAdls))
-	}
-	return mountOptions
-}
-
-func getStreamingMountOptions(streaming *radixv1.BlobFuse2StreamingOptions) []string {
-	var mountOptions []string
-	if streaming != nil && streaming.Enabled != nil && !*streaming.Enabled {
-		return nil
-	}
-	mountOptions = append(mountOptions, "--streaming=true")
-
-	var streamCache uint64 = 750
-	if streaming != nil && streaming.StreamCache != nil {
-		streamCache = *streaming.StreamCache
-	}
-	mountOptions = append(mountOptions, fmt.Sprintf("--block-cache-pool-size=%v", streamCache))
-	return mountOptions
-}
-
-func getVolumeMountAccessMode(radixVolumeMount *radixv1.RadixVolumeMount) corev1.PersistentVolumeAccessMode {
-	//nolint:staticcheck
-	accessMode := radixVolumeMount.AccessMode
-	if radixVolumeMount.BlobFuse2 != nil {
-		accessMode = radixVolumeMount.BlobFuse2.AccessMode
-	}
-	switch strings.ToLower(accessMode) {
-	case strings.ToLower(string(corev1.ReadWriteOnce)):
-		return corev1.ReadWriteOnce
-	case strings.ToLower(string(corev1.ReadWriteMany)):
-		return corev1.ReadWriteMany
-	case strings.ToLower(string(corev1.ReadWriteOncePod)):
-		return corev1.ReadWriteOncePod
-	}
-	return corev1.ReadOnlyMany // default access mode
-}
-
-func getRadixBlobFuse2VolumeMountUid(radixVolumeMount *radixv1.RadixVolumeMount) string {
-	if radixVolumeMount.BlobFuse2 != nil {
-		return radixVolumeMount.BlobFuse2.UID
-	}
-	//nolint:staticcheck
-	return radixVolumeMount.UID
-}
-
-func getRadixBlobFuse2VolumeMountGid(radixVolumeMount *radixv1.RadixVolumeMount) string {
-	if radixVolumeMount.BlobFuse2 != nil {
-		return radixVolumeMount.BlobFuse2.GID
-	}
-	//nolint:staticcheck
-	return radixVolumeMount.GID
-}
-
-func getRadixBlobFuse2VolumeMountContainerName(radixVolumeMount *radixv1.RadixVolumeMount) string {
-	if radixVolumeMount.BlobFuse2 != nil {
-		return radixVolumeMount.BlobFuse2.Container
-	}
-	//nolint:staticcheck
-	return radixVolumeMount.Storage
-}
-
-func getRadixBlobFuse2VolumeMountRequestsStorage(radixVolumeMount *radixv1.RadixVolumeMount) string {
-	if radixVolumeMount.BlobFuse2 != nil {
-		return radixVolumeMount.BlobFuse2.RequestsStorage
-	}
-	//nolint:staticcheck
-	return radixVolumeMount.RequestsStorage
 }
 
 func deletePvc(ctx context.Context, kubeClient kubernetes.Interface, namespace, pvcName string) error {
@@ -706,15 +473,6 @@ func deletePv(ctx context.Context, kubeClient kubernetes.Interface, pvName strin
 		return err
 	}
 	return nil
-}
-
-func getRadixVolumeMountStorage(radixVolumeMount *radixv1.RadixVolumeMount) string {
-	blobFuse2VolumeMountContainer := getRadixBlobFuse2VolumeMountContainerName(radixVolumeMount)
-	if len(blobFuse2VolumeMountContainer) != 0 {
-		return blobFuse2VolumeMountContainer
-	}
-	//nolint:staticcheck
-	return radixVolumeMount.Storage
 }
 
 func garbageCollectOrphanedCsiAzurePvs(ctx context.Context, kubeClient kubernetes.Interface, excludePvcNames map[string]any) error {
@@ -747,125 +505,6 @@ func garbageCollectOrphanedCsiAzurePvs(ctx context.Context, kubeClient kubernete
 func knownCSIDriver(driver string) bool {
 	_, ok := csiVolumeProvisioners[driver]
 	return ok
-}
-
-func createOrUpdateCsiAzureVolumeResourcesForVolumes(ctx context.Context, kubeClient kubernetes.Interface, radixDeployment *radixv1.RadixDeployment, namespace, componentName string, identity *radixv1.Identity, desiredVolumes []corev1.Volume) ([]corev1.Volume, error) {
-	functionalPvList, err := getCsiAzurePvsForNamespace(ctx, kubeClient, namespace, true)
-	if err != nil {
-		return nil, err
-	}
-	pvcByNameMap, err := getComponentPvcByNameMap(ctx, kubeClient, namespace, componentName)
-	if err != nil {
-		return nil, err
-	}
-	radixVolumeMountsByNameMap := getRadixVolumeMountsByNameMap(radixDeployment, componentName)
-	var errs []error
-	var volumes []corev1.Volume
-	for _, volume := range desiredVolumes {
-		if volume.PersistentVolumeClaim == nil {
-			volumes = append(volumes, volume)
-			continue
-		}
-		radixVolumeMount, existsRadixVolumeMount := radixVolumeMountsByNameMap[volume.Name]
-		if !existsRadixVolumeMount {
-			continue
-		}
-		processedVolume, err := createOrUpdateCsiAzureVolumeResourcesForVolume(ctx, kubeClient, radixDeployment, namespace, componentName, identity, volume, radixVolumeMount, functionalPvList, pvcByNameMap)
-		if err != nil {
-			errs = append(errs, err)
-			continue
-		}
-		if processedVolume != nil {
-			volumes = append(volumes, *processedVolume)
-		}
-	}
-	if len(errs) > 0 {
-		return nil, errors.Join(errs...)
-	}
-	return volumes, nil
-}
-
-func createOrUpdateCsiAzureVolumeResourcesForVolume(ctx context.Context, kubeClient kubernetes.Interface, radixDeployment *radixv1.RadixDeployment, namespace string, componentName string, identity *radixv1.Identity, volume corev1.Volume, radixVolumeMount *radixv1.RadixVolumeMount, persistentVolumes []corev1.PersistentVolume, pvcByNameMap map[string]*corev1.PersistentVolumeClaim) (*corev1.Volume, error) {
-	if volume.PersistentVolumeClaim == nil {
-		return &volume, nil
-	}
-	appName := radixDeployment.Spec.AppName
-	pvcName := volume.PersistentVolumeClaim.ClaimName
-	existingPvc := pvcByNameMap[pvcName]
-
-	pvName := getCsiAzurePvName()
-	existingPv, pvExists := getCsiAzureComponentPvByPvcName(persistentVolumes, pvcName)
-	if pvExists {
-		radixVolumeMountPv := buildCsiAzurePv(appName, namespace, componentName, existingPv.GetName(), pvcName, radixVolumeMount, identity)
-		if !matchComponentVolumeMountLabels(namespace, appName, componentName, radixVolumeMount, existingPv, existingPvc) ||
-			!EqualPersistentVolumes(existingPv, radixVolumeMountPv) {
-			pvExists = false
-			newPvcName, err := getCsiAzurePvcName(componentName, radixVolumeMount)
-			if err != nil {
-				return nil, err
-			}
-			pvcName = newPvcName
-		}
-	} else {
-		existingPv, pvExists = getCsiAzureComponentPvByRadixVolumeMountContent(namespace, appName, componentName, radixVolumeMount, persistentVolumes, pvcName, existingPvc, identity)
-	}
-	if pvExists {
-		pvName = existingPv.GetName()
-		pvcName = existingPv.Spec.ClaimRef.Name
-	}
-
-	existingPvc, pvcExists := pvcByNameMap[pvcName]
-	if !pvExists && pvcExists {
-		pvName = existingPvc.Spec.VolumeName
-		if len(pvName) == 0 {
-			pvName = getCsiAzurePvName() // for auto-provisioned persistent volume
-		}
-	}
-	needToCreatePvc := !pvcExists
-	needToReCreatePv := pvExists && !pvcExists && len(existingPv.Spec.StorageClassName) > 0 // HACK: always re-create PV if it uses SC and PVC is missing
-	radixVolumeMountPvc := buildPvc(appName, namespace, componentName, pvName, pvcName, radixVolumeMount)
-	needToReCreatePvc := pvcExists && (existingPvc.Spec.VolumeName != pvName || !EqualPersistentVolumeClaims(existingPvc, radixVolumeMountPvc))
-
-	if needToReCreatePv || needToReCreatePvc {
-		newPvcName, err := getCsiAzurePvcName(componentName, radixVolumeMount)
-		if err != nil {
-			return nil, err
-		}
-		pvcName = newPvcName
-
-		pvExists = false
-		pvName = getCsiAzurePvName()
-	}
-
-	if !pvExists || needToReCreatePv {
-		log.Ctx(ctx).Debug().Msgf("Create PersistentVolume %s in namespace %s", pvName, namespace)
-		pv := buildCsiAzurePv(appName, namespace, componentName, pvName, pvcName, radixVolumeMount, identity)
-		if _, err := kubeClient.CoreV1().PersistentVolumes().Create(ctx, pv, metav1.CreateOptions{}); err != nil {
-			return nil, err
-		}
-	}
-
-	if needToCreatePvc || needToReCreatePvc {
-		radixVolumeMountPvc.SetName(pvcName)
-		radixVolumeMountPvc.Spec.VolumeName = pvName
-		log.Ctx(ctx).Debug().Msgf("Create PersistentVolumeClaim %s in namespace %s for PersistentVolume %s", radixVolumeMountPvc.GetName(), namespace, pvName)
-		if _, err := kubeClient.CoreV1().PersistentVolumeClaims(namespace).Create(ctx, radixVolumeMountPvc, metav1.CreateOptions{}); err != nil {
-			return nil, err
-		}
-	}
-	volume.PersistentVolumeClaim.ClaimName = pvcName // in case it was updated with new name
-	return &volume, nil
-}
-
-func getComponentPvcByNameMap(ctx context.Context, kubeClient kubernetes.Interface, namespace string, componentName string) (map[string]*corev1.PersistentVolumeClaim, error) {
-	pvcList, err := getCsiAzurePvcs(ctx, kubeClient, namespace)
-	if err != nil {
-		return nil, err
-	}
-	pvcs := slice.FindAll(pvcList.Items, func(pvc corev1.PersistentVolumeClaim) bool {
-		return pvc.GetLabels()[kube.RadixComponentLabel] == componentName
-	})
-	return GetPersistentVolumeClaimMap(&pvcs), nil
 }
 
 func getCurrentlyUsedPvcNames(ctx context.Context, kubeClient kubernetes.Interface, radixDeployment *radixv1.RadixDeployment) (map[string]any, error) {
@@ -940,83 +579,6 @@ func garbageCollectCsiAzurePvcs(ctx context.Context, kubeClient kubernetes.Inter
 	return nil
 }
 
-func getCsiAzureComponentPvByPvcName(persistentVolumes []corev1.PersistentVolume, pvcName string) (*corev1.PersistentVolume, bool) {
-	pv, ok := slice.FindFirst(persistentVolumes, func(pv corev1.PersistentVolume) bool {
-		return pv.Spec.PersistentVolumeSource.CSI != nil && pv.Spec.ClaimRef != nil && pv.Spec.ClaimRef.Name == pvcName
-	})
-	return &pv, ok
-}
-
-func getCsiAzureComponentPvByRadixVolumeMountContent(appName string, namespace string, componentName string, radixVolumeMount *radixv1.RadixVolumeMount, persistentVolumes []corev1.PersistentVolume, pvcName string, pvc *corev1.PersistentVolumeClaim, identity *radixv1.Identity) (*corev1.PersistentVolume, bool) {
-	for _, pv := range persistentVolumes {
-		if pv.Spec.PersistentVolumeSource.CSI == nil ||
-			pv.Spec.ClaimRef == nil ||
-			!matchComponentVolumeMountLabels(namespace, appName, componentName, radixVolumeMount, &pv, pvc) {
-			continue
-		}
-		radixVolumeMountPv := buildCsiAzurePv(appName, namespace, componentName, pv.GetName(), pvcName, radixVolumeMount, identity)
-		if EqualPersistentVolumes(&pv, radixVolumeMountPv) {
-			return &pv, true
-		}
-	}
-	return nil, false
-}
-
-func matchComponentVolumeMountLabels(namespace, appName, componentName string, radixVolumeMount *radixv1.RadixVolumeMount, pv *corev1.PersistentVolume, pvc *corev1.PersistentVolumeClaim) bool {
-	pvLabels := pv.GetLabels()
-	if _, ok := pvLabels[kube.RadixAppLabel]; ok && matchComponentVolumeMountPvLabels(pvLabels, namespace, appName, componentName, radixVolumeMount) {
-		return true
-	}
-	return pvc != nil && matchComponentVolumeMountPvcLabels(pvc.GetLabels(), appName, componentName, radixVolumeMount)
-}
-
-func matchComponentVolumeMountPvLabels(labels map[string]string, namespace, appName, componentName string, radixVolumeMount *radixv1.RadixVolumeMount) bool {
-	return labels[kube.RadixNamespace] == namespace &&
-		labels[kube.RadixAppLabel] == appName &&
-		labels[kube.RadixComponentLabel] == componentName &&
-		labels[kube.RadixVolumeMountNameLabel] == radixVolumeMount.Name
-}
-
-func matchComponentVolumeMountPvcLabels(labels map[string]string, appName, componentName string, radixVolumeMount *radixv1.RadixVolumeMount) bool {
-	return labels[kube.RadixAppLabel] == appName &&
-		labels[kube.RadixComponentLabel] == componentName &&
-		labels[kube.RadixVolumeMountNameLabel] == radixVolumeMount.Name &&
-		labels[kube.RadixMountTypeLabel] == string(GetCsiAzureVolumeMountType(radixVolumeMount))
-}
-
-func getRadixVolumeMountsByNameMap(radixDeployment *radixv1.RadixDeployment, componentName string) map[string]*radixv1.RadixVolumeMount {
-	volumeMountsByNameMap := make(map[string]*radixv1.RadixVolumeMount)
-	for _, component := range radixDeployment.Spec.Components {
-		if findCsiAzureVolumeForComponent(volumeMountsByNameMap, component.VolumeMounts, componentName, &component) {
-			break
-		}
-	}
-	for _, component := range radixDeployment.Spec.Jobs {
-		if findCsiAzureVolumeForComponent(volumeMountsByNameMap, component.VolumeMounts, componentName, &component) {
-			break
-		}
-	}
-	return volumeMountsByNameMap
-}
-
-func findCsiAzureVolumeForComponent(volumeMountsByNameMap map[string]*radixv1.RadixVolumeMount, volumeMounts []radixv1.RadixVolumeMount, componentName string, component radixv1.RadixCommonDeployComponent) bool {
-	if !strings.EqualFold(componentName, component.GetName()) {
-		return false
-	}
-	for _, radixVolumeMount := range volumeMounts {
-		if radixVolumeMount.BlobFuse2 == nil && !isKnownCsiAzureVolumeMount(string(GetCsiAzureVolumeMountType(&radixVolumeMount))) {
-			continue
-		}
-		radixVolumeMount := radixVolumeMount
-		volumeMountName, err := getCsiAzureVolumeMountName(componentName, &radixVolumeMount)
-		if err != nil {
-			return false
-		}
-		volumeMountsByNameMap[volumeMountName] = &radixVolumeMount
-	}
-	return true
-}
-
 func trimVolumeNameToValidLength(volumeName string) string {
 	if len(volumeName) <= volumeNameMaxLength {
 		return volumeName
@@ -1061,17 +623,4 @@ func garbageCollectSecrets(ctx context.Context, kubeUtil *kube.Kube, namespace s
 		}
 	}
 	return nil
-}
-
-// GetExistingJobAuxComponentVolumes Get existing job aux component volumes
-func GetExistingJobAuxComponentVolumes(ctx context.Context, kubeUtil *kube.Kube, namespace, jobComponentName string) ([]corev1.Volume, error) {
-	jobAuxKubeDeploymentName := defaults.GetJobAuxKubeDeployName(jobComponentName)
-	jobAuxKubeDeployment, err := kubeUtil.KubeClient().AppsV1().Deployments(namespace).Get(ctx, jobAuxKubeDeploymentName, metav1.GetOptions{})
-	if err != nil {
-		if k8serrors.IsNotFound(err) {
-			return nil, nil
-		}
-		return nil, err
-	}
-	return jobAuxKubeDeployment.Spec.Template.Spec.Volumes, nil
 }
